# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoBird Counter API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type AnalysisJobStatus =
  (typeof AnalysisJobStatus)[keyof typeof AnalysisJobStatus];

export const AnalysisJobStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface AnalysisJob {
  id: string;
  status: AnalysisJobStatus;
  /**
   * @minimum 0
   * @maximum 100
   */
  progress: number;
  filename: string;
  sampleInterval: number;
  createdAt: string;
  completedAt?: string | null;
  totalFrames?: number | null;
  processedFrames?: number | null;
}

export interface SpeciesDetection {
  species: string;
  totalCount: number;
  averageConfidence: number;
  color: string;
  lastSeenAt?: string | null;
}

export interface DetectionSummary {
  totalBirdsDetected: number;
  uniqueSpecies: number;
  mostCommonSpecies?: string | null;
  processingDurationSeconds?: number | null;
}

export interface AnalysisResult {
  job: AnalysisJob;
  detections: SpeciesDetection[];
  summary?: DetectionSummary | null;
}

export interface BirdDetection {
  trackId: number;
  species: string;
  confidence: number;
  /**
   * @minItems 4
   * @maxItems 4
   */
  bbox: number[];
  color: string;
}

export interface DetectionFrame {
  frameIndex: number;
  timestamp: number;
  detections: BirdDetection[];
}

export type SpeciesInfoCardSource =
  (typeof SpeciesInfoCardSource)[keyof typeof SpeciesInfoCardSource];

export const SpeciesInfoCardSource = {
  ai: "ai",
  cache: "cache",
} as const;

export interface SpeciesInfoCard {
  speciesName: string;
  scientificName: string;
  conservationStatus: string;
  habitat: string;
  diet: string;
  funFact: string;
  description: string;
  ebirdOccurrences?: string | null;
  source: SpeciesInfoCardSource;
}

export type SpeciesStatsTopSpeciesItem = {
  species: string;
  count: number;
};

export type SpeciesStatsRecentActivityItem = {
  date: string;
  detections: number;
};

export interface SpeciesStats {
  totalAnalyses: number;
  totalDetections: number;
  topSpecies: SpeciesStatsTopSpeciesItem[];
  recentActivity: SpeciesStatsRecentActivityItem[];
}

export type UploadVideoBody = {
  video: Blob;
  /**
   * Sample interval in seconds (0.5 to 5.0)
   * @minimum 0.5
   * @maximum 5
   */
  sampleInterval?: number;
};