# lib/api-zod/src/generated/types/analysisJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoBird Counter API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisJobStatus } from "./analysisJobStatus";

export interface AnalysisJob {
  id: string;
  status: AnalysisJobStatus;
  /**
   * @minimum 0
   * @maximum 100
   */
  progress: number;
  filename: string;
  sampleInterval: number;
  createdAt: Date;
  completedAt?: Date | null;
  totalFrames?: number | null;
  processedFrames?: number | null;
}