# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoBird Counter API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisJob } from "./analysisJob";
import type { DetectionSummary } from "./detectionSummary";
import type { SpeciesDetection } from "./speciesDetection";

export interface AnalysisResult {
  job: AnalysisJob;
  detections: SpeciesDetection[];
  summary?: DetectionSummary | null;
}