# lib/api-zod/src/generated/types/birdDetection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoBird Counter API
 * OpenAPI spec version: 0.1.0
 */

export interface BirdDetection {
  trackId: number;
  species: string;
  confidence: number;
  /**
   * @minItems 4
   * @maxItems 4
   */
  bbox: number[];
  color: string;
}