# lib/api-zod/src/generated/types/detectionFrame.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoBird Counter API
 * OpenAPI spec version: 0.1.0
 */
import type { BirdDetection } from "./birdDetection";

export interface DetectionFrame {
  frameIndex: number;
  timestamp: number;
  detections: BirdDetection[];
}