# lib/api-zod/src/generated/types/detectionSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoBird Counter API
 * OpenAPI spec version: 0.1.0
 */

export interface DetectionSummary {
  totalBirdsDetected: number;
  uniqueSpecies: number;
  mostCommonSpecies?: string | null;
  processingDurationSeconds?: number | null;
}