# lib/api-zod/src/generated/types/speciesDetection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoBird Counter API
 * OpenAPI spec version: 0.1.0
 */

export interface SpeciesDetection {
  species: string;
  totalCount: number;
  averageConfidence: number;
  color: string;
  lastSeenAt?: Date | null;
}