# lib/api-zod/src/generated/types/speciesInfoCard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoBird Counter API
 * OpenAPI spec version: 0.1.0
 */
import type { SpeciesInfoCardSource } from "./speciesInfoCardSource";

export interface SpeciesInfoCard {
  speciesName: string;
  scientificName: string;
  conservationStatus: string;
  habitat: string;
  diet: string;
  funFact: string;
  description: string;
  ebirdOccurrences?: string | null;
  source: SpeciesInfoCardSource;
}