# lib/api-zod/src/generated/types/speciesStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoBird Counter API
 * OpenAPI spec version: 0.1.0
 */
import type { SpeciesStatsRecentActivityItem } from "./speciesStatsRecentActivityItem";
import type { SpeciesStatsTopSpeciesItem } from "./speciesStatsTopSpeciesItem";

export interface SpeciesStats {
  totalAnalyses: number;
  totalDetections: number;
  topSpecies: SpeciesStatsTopSpeciesItem[];
  recentActivity: SpeciesStatsRecentActivityItem[];
}